fix: 認証ページの共通部分をまとめる
